requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.4
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.4
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
